requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.3.0
beir==2.0.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.40.0
datasets==3.2.0
Deprecated==1.2.15
dill==0.3.8
docker==7.1.0
elasticsearch==7.9.1
faiss-cpu==1.9.0.post1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
levenshtein>=0.26.1
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nlpaug==1.1.11
nltk>=3.9.1
numpy>=1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytrec_eval==0.5
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge-score>=0.1.2
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.0
sentence-transformers==3.3.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tensorflow-hub>=0.16.1
textattack==0.3.10
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchmetrics>=1.6.1
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
nltk==3.9.1
wordnet==0.0.1b2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
spacy==3.8.4
